Lexer+Parser+Codegen: Replace semicolons with EOL tokens
This patch removes semicolons where possibile. Infact in for-statement
this are still required to split the initializer from the condition.
Moreover from this patch and so on you should not be allowed anymore to
have more statements on the same line unfortunately. Though this change
was made primarly to improve error reporting dramatically. :^)