Parser+Codegen: Improve match statemente expression handling
We are now able to parse all kinds of expression inside of a match
statement. Note that this required one major change: we now have to
switch to a C++ backend, which i am totally fine with. The other thing
is that we have to specify the enum while writing the match cases